# swagger/swagger.yaml
basePath: /v1
definitions:
  error_handler.ErrorResponse:
    properties:
      code:
        type: integer
      details:
        additionalProperties:
          type: string
        type: object
      error:
        type: string
    type: object
...
  model.Steam2FARequest:
    properties:
      completedAt:
        type: string
      errorMsg:
        type: string
      id:
        type: string
      message:
        type: string
      requestTime:
        type: string
      serverId:
        type: string
      status:
        $ref: '#/definitions/model.Steam2FAStatus'
    type: object
  model.Steam2FAStatus:
    enum:
    - idle
    - pending
    - complete
    - error
    type: string
    x-enum-varnames:
    - Steam2FAStatusIdle
    - Steam2FAStatusPending
    - Steam2FAStatusComplete
    - Steam2FAStatusError
...
host: acc-api.jurmanovic.com
info:
  contact:
    name: ACC Server Manager Support
    url: https://github.com/FJurmanovic/acc-server-manager
  description: API for managing Assetto Corsa Competizione dedicated servers
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: ACC Server Manager API
  version: "1.0"
paths:
...
  /steam2fa/{id}:
    get:
      consumes:
      - application/json
      description: Get a specific Steam 2FA authentication request by ID
      parameters:
      - description: 2FA Request ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Steam2FARequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
      summary: Get 2FA request
      tags:
      - Steam 2FA
  /steam2fa/{id}/cancel:
    post:
      consumes:
      - application/json
      description: Cancel a Steam 2FA authentication request
      parameters:
      - description: 2FA Request ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Steam2FARequest'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
      summary: Cancel 2FA request
      tags:
      - Steam 2FA
  /steam2fa/{id}/complete:
    post:
      consumes:
      - application/json
      description: Mark a Steam 2FA authentication request as completed
      parameters:
      - description: 2FA Request ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Steam2FARequest'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
      summary: Complete 2FA request
      tags:
      - Steam 2FA
  /steam2fa/pending:
    get:
      consumes:
      - application/json
      description: Get all pending Steam 2FA authentication requests
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Steam2FARequest'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
      summary: Get pending 2FA requests
      tags:
      - Steam 2FA
...
schemes:
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"